backend/requirements.txt
fastapi==0.115.11
pydantic==2.10.6
pytest==8.3.5
SQLAlchemy==2.0.38
uvicorn==0.34.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
openai==1.66.3
azure-data-tables==12.6.0
openai==1.66.3
azure-servicebus==7.14.1
